Move get_default and get_dtype to ObjectNode
